+++ src/Utility/DatabaseUtility.php
@@ -2,18 +2,9 @@
 
 namespace SourceBroker\DeployerExtendedDatabase\Utility;
 
-/**
- * Class DatabaseUtility
- * @package SourceBroker\DeployerExtendedDatabase\Utility
- */
 class DatabaseUtility
 {
-
-    /**
-     * @param $dbConf
-     * @return array
-     */
-    public function getTables($dbConf): array
+    public function getTables(array $dbConf): array
     {
         $link = mysqli_connect($dbConf['host'], $dbConf['user'], $dbConf['password'], $dbConf['dbname'],
             $dbConf['port']);